+++ tests/testthat/test-mvn.R
@@ -23,3 +23,17 @@ test_that("Generate MVN data w/ interventions", {
     expect_equal(as.vector(dat[, 1:nivn]), rep(1, nobs*nivn)) # first three columns are all ones
 })
 
+test_that("ivn.rand = FALSE with invalid input returns an error (fixes issue #4)", {
+    nobs <- 10
+    nivn <- 3
+
+    ### character / node names input
+    ivn <- lapply(1:nobs, function(x) names(el)[1:nivn]) # only intervene on first 3 nodes
+    expect_error(generate_mvn_data(el, params, n = nobs, ivn = ivn, ivn.rand = FALSE),
+                 "ivn.rand set to FALSE with invalid input")
+
+    ### numeric / index input
+    ivn <- lapply(1:nobs, function(x) 1:nivn) # only intervene on first 3 nodes
+    expect_error(generate_mvn_data(el, params, n = nobs, ivn = ivn, ivn.rand = FALSE),
+                 "ivn.rand set to FALSE with invalid input")
+})